+++ processor/src/main/java/react4j/processor/Generator.java
@@ -31,7 +31,6 @@ import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
 import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactElement;
 import react4j.core.ReactNode;
 
 final class Generator
@@ -79,7 +78,7 @@ final class Generator
       builder.addField( buildEventHandlerField( eventHandler ).build() );
     }
 
-    builder.addMethod( buildFactoryMethod( descriptor ).build() );
+    builder.addMethod( buildFactoryMethod().build() );
     builder.addMethod( buildFactory2Method( descriptor ).build() );
     builder.addMethod( buildFactory3Method( descriptor ).build() );
 
@@ -129,14 +128,6 @@ final class Generator
     return ParameterizedTypeName.get( ClassName.get( ComponentConstructorFunction.class ),
                                       TypeName.get( descriptor.getPropsType().asType() ),
                                       TypeName.get( descriptor.getStateType().asType() ),
-                                      ClassName.bestGuess( "NativeReactComponent" ) );
-  }
-
-  @Nonnull
-  private static ParameterizedTypeName getReactElementType( @Nonnull final ComponentDescriptor descriptor )
-  {
-    return ParameterizedTypeName.get( ClassName.get( ReactElement.class ),
-                                      TypeName.get( descriptor.getPropsType().asType() ),
                                       ClassName.bestGuess( "NativeReactComponent" ) );
   }
 
@@ -267,12 +258,12 @@ final class Generator
   }
 
   @Nonnull
-  private static MethodSpec.Builder buildFactoryMethod( @Nonnull final ComponentDescriptor descriptor )
+  private static MethodSpec.Builder buildFactoryMethod()
   {
     return MethodSpec.methodBuilder( "_create" ).
       addAnnotation( Nonnull.class ).
       addModifiers( Modifier.STATIC ).
-      returns( getReactElementType( descriptor ) ).
+      returns( ClassName.get( ReactNode.class ) ).
       addStatement( "return $T.createElement( TYPE )", React.class );
   }
 
@@ -282,7 +273,7 @@ final class Generator
     return MethodSpec.methodBuilder( "_create" ).
       addAnnotation( Nonnull.class ).
       addModifiers( Modifier.STATIC ).
-      returns( getReactElementType( descriptor ) ).
+      returns( ClassName.get( ReactNode.class ) ).
       addParameter( ParameterSpec.builder( ClassName.get( descriptor.getPropsType() ), "props", Modifier.FINAL ).
         addAnnotation( Nullable.class ).build() ).
       addStatement( "return $T.createElement( TYPE, props )", React.class );
@@ -294,7 +285,7 @@ final class Generator
     return MethodSpec.methodBuilder( "_create" ).
       addAnnotation( Nonnull.class ).
       addModifiers( Modifier.STATIC ).
-      returns( getReactElementType( descriptor ) ).
+      returns( ClassName.get( ReactNode.class ) ).
       addParameter( ParameterSpec.builder( ClassName.get( descriptor.getPropsType() ), "props", Modifier.FINAL ).
         addAnnotation( Nullable.class ).build() ).
       addParameter( ParameterSpec.builder( ReactNode.class, "child", Modifier.FINAL ).