+++ processor/src/test/resources/expected/com/example/arez/ComponentWithEventHandler_.java
@@ -16,7 +16,6 @@ import react4j.core.NativeAdapterComponent;
 import react4j.core.Procedure;
 import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactElement;
 import react4j.core.ReactNode;
 
 @ArezComponent(
@@ -33,17 +32,17 @@ class ComponentWithEventHandler_ extends ComponentWithEventHandler {
   private final ComponentWithEventHandler.CustomHandler _handleFoo2 = create_handleFoo2();
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> _create() {
+  static ReactNode _create() {
     return React.createElement( TYPE );
   }
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props) {
+  static ReactNode _create(@Nullable final BaseProps props) {
     return React.createElement( TYPE, props );
   }
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
+  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
     return React.createElement( TYPE, props, child );
   }