+++ processor/src/test/resources/expected/com/example/basic/CustomPropsAndStateReactComponent_.java
@@ -8,7 +8,6 @@ import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
 import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactElement;
 import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
@@ -16,17 +15,17 @@ class CustomPropsAndStateReactComponent_ extends CustomPropsAndStateReactCompone
   private static final ComponentConstructorFunction<CustomPropsAndStateReactComponent.Props, CustomPropsAndStateReactComponent.State, NativeReactComponent> TYPE = getConstructorFunction();
 
   @Nonnull
-  static ReactElement<CustomPropsAndStateReactComponent.Props, NativeReactComponent> _create() {
+  static ReactNode _create() {
     return React.createElement( TYPE );
   }
 
   @Nonnull
-  static ReactElement<CustomPropsAndStateReactComponent.Props, NativeReactComponent> _create(@Nullable final CustomPropsAndStateReactComponent.Props props) {
+  static ReactNode _create(@Nullable final CustomPropsAndStateReactComponent.Props props) {
     return React.createElement( TYPE, props );
   }
 
   @Nonnull
-  static ReactElement<CustomPropsAndStateReactComponent.Props, NativeReactComponent> _create(@Nullable final CustomPropsAndStateReactComponent.Props props, @Nullable final ReactNode child) {
+  static ReactNode _create(@Nullable final CustomPropsAndStateReactComponent.Props props, @Nullable final ReactNode child) {
     return React.createElement( TYPE, props, child );
   }