+++ processor/src/test/resources/expected/com/example/event_handler/CustomNameComponent_.java
@@ -13,7 +13,6 @@ import react4j.core.NativeAdapterComponent;
 import react4j.core.Procedure;
 import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactElement;
 import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
@@ -24,17 +23,17 @@ class CustomNameComponent_ extends CustomNameComponent {
   private final Procedure _handleFoo = create_handleFoo();
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> _create() {
+  static ReactNode _create() {
     return React.createElement( TYPE );
   }
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props) {
+  static ReactNode _create(@Nullable final BaseProps props) {
     return React.createElement( TYPE, props );
   }
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
+  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
     return React.createElement( TYPE, props, child );
   }