+++ todomvc/src/main/java/react4j/todomvc/Footer.java
@@ -8,7 +8,6 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
-import react4j.core.ReactElement;
 import react4j.core.ReactNode;
 import react4j.dom.events.MouseEventHandler;
 import react4j.dom.proptypes.html.AnchorProps;
@@ -30,7 +29,7 @@ class Footer
   }
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> create()
+  static ReactNode create()
   {
     return _create();
   }