+++ todomvc/src/main/java/react4j/todomvc/TodoEntry.java
@@ -12,7 +12,6 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
-import react4j.core.ReactElement;
 import react4j.core.ReactNode;
 import react4j.dom.events.FormEvent;
 import react4j.dom.events.FormEventHandler;
@@ -43,7 +42,7 @@ class TodoEntry
   }
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> create()
+  static ReactNode create()
   {
     return _create();
   }