+++ todomvc/src/main/java/react4j/todomvc/TodoList.java
@@ -11,7 +11,6 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
-import react4j.core.ReactElement;
 import react4j.core.ReactNode;
 import react4j.dom.events.FormEvent;
 import react4j.dom.events.FormEventHandler;
@@ -41,7 +40,7 @@ class TodoList
   }
 
   @Nonnull
-  static ReactElement<BaseProps, NativeReactComponent> create()
+  static ReactNode create()
   {
     return _create();
   }